backend/requirements.txt
flask
openai
python-dotenv
pymongo